requirements.txt
flask
line-bot-sdk
wheel
grpcio
google-cloud-firestore
flask_cors
line-bot-sdk
google-cloud-storage
pandas
gunicorn
google-cloud-logging
tensorflow-cpu == 2.4.1
opencv-python==4.5.3.56
Pillow
pygsheets
gspread
google-auth
oauth2client
tabulate